+++ test/shopify-cli/theme/dev_server/integration_test.rb
@@ -196,6 +196,9 @@ module ShopifyCLI
             .to_return(status: 200, body: "{}")
           stub_request(:any, THEMES_API_URL)
             .to_return(status: 200, body: "{}")
+          # Stub request to get deleted file after file.delete is called
+          stub_request(:get, "#{ASSETS_API_URL}?asset%5Bkey%5D=assets/added.css")
+            .to_return(status: 200, body: "{}")
 
           start_server
           # Wait for server to start & sync the files
@@ -212,7 +215,7 @@ module ShopifyCLI
           end
         end
 
-        def with_retries(*exceptions, retries: 5)
+        def with_retries(*exceptions, retries: 10)
           yield
         rescue *exceptions
           retries -= 1